XrHand.getRadius > XrHand.radius, moved XrFinger and XrJoint to own files
